[REFACTOR] Regroup auth module
